Revert "Refactor/validator"

+++ src/presentation/controller/delete-beer-style-controller.ts
@@ -1,12 +1,15 @@
 import { IDeleteBeerStyleService } from "@/application/interfaces";
-import { serverError, ok, notFound } from "@/presentation/helpers";
+import { serverError, ok, badRequest, notFound } from "@/presentation/helpers";
 import { IController, IHttpRequest, IHttpResponse } from "@/presentation/interfaces";
+import { MissinParamError } from "../errors";
 
 export class DeleteBeerStyleController implements IController {
   constructor(private readonly beerStyleService: IDeleteBeerStyleService) {}
 
   async handle(httpRequest: IHttpRequest): Promise<IHttpResponse> {
     try {
+      if (!httpRequest.params?.id) return badRequest(new MissinParamError("id"));
+
       const { id } = httpRequest.params;
 
       const beerStyle = await this.beerStyleService.delete(id);
